pyproject.toml
[tool.poetry]
name = "severstal-fastapi"
version = "0.1.0"
description = ""
authors = ["Anna <pompeeva.anna@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
